CLOUD-61407 dash-hdfs.sh and dash-local.sh scripts was changed to be idempotent

+++ core/src/main/resources/scripts/dash-hdfs.sh
@@ -27,32 +27,43 @@ main(){
 
   cd /tmp
   for tar_file_path in $MR_TAR_PATH $TEZ_TAR_PATH; do
+    if hadoop fs -test -e $tar_file_path; then
+      tar_file=$(basename $tar_file_path)
 
-    tar_file=$(basename $tar_file_path)
+      echo "Copying $tar_file from HDFS to local fs."
+      [[ -f "/tmp/$tar_file" ]] && rm "/tmp/$tar_file"
+      hadoop fs -copyToLocal "$tar_file_path" /tmp
 
-    echo "Copying $tar_file from HDFS to local fs."
-    [[ -f "/tmp/$tar_file" ]] && rm "/tmp/$tar_file"
-    hadoop fs -copyToLocal "$tar_file_path" /tmp
+      rm -rf $TMP_EXTRACT_DIR && mkdir $TMP_EXTRACT_DIR && cd $TMP_EXTRACT_DIR
+      echo "Extracting $tar_file."
+      tar -zxf ../$tar_file
 
-    rm -rf $TMP_EXTRACT_DIR && mkdir $TMP_EXTRACT_DIR && cd $TMP_EXTRACT_DIR
-    echo "Extracting $tar_file."
-    tar -zxf ../$tar_file
+      jarfiles=( $(find . -name "azure-storage*.jar") )
+      if [ ${#jarfiles[@]} -gt 0 ]; then
+        echo "Replacing azure-storage.jar with $STORAGE_JAR in $tar_file."
+        for jarfile in ${jarfiles[@]}; do
+          echo "Replacing $jarfile"
+          cp -f $SOURCE_JAR ${jarfile%azure*}
+          rm -f $jarfile
+        done
+        echo "Removing $tar_file."
+        rm -f "../$tar_file"
 
-    echo "Replacing azure-storage.jar with $STORAGE_JAR."
-    find . -name "azure-storage*.jar" | while read line; do echo Replace $line; \cp -f $SOURCE_JAR ${line%azure*}; rm -f $line; done
+        echo "Creating new $tar_file with the replaced libs."
+        tar -zcf "../$tar_file" $(ls)
 
-    echo "Removing $tar_file."
-    rm -f "../$tar_file"
+        echo "Replacing $tar_file on HDFS."
+        su -c "hadoop fs -put -f ../$tar_file $tar_file_path" hdfs
+      else
+        echo "Replacing azure-storage.jar with $STORAGE_JAR already done in $tar_file on HDFS, skip it."
+      fi
 
-    echo "Creating new $tar_file with the replaced libs."
-    tar -zcf "../$tar_file" $(ls)
-
-    echo "Replacing $tar_file on HDFS."
-    su -c "hadoop fs -put -f ../$tar_file $tar_file_path" hdfs
-
-    echo "Cleaning up local fs."
-    rm -f "../$tar_file"
-    cd .. && rm -rf $TM_EXTRACT_DIR
+      echo "Cleaning up local fs."
+      rm -f "../$tar_file"
+      cd .. && rm -rf $TM_EXTRACT_DIR
+    else
+      echo "$tar_file_path does not exist!"
+    fi
   done
 }
